@@ -29,10 +29,8 @@ class RouteServiceProvider extends ServiceProvider
      *
      * @return void
      */
-    public function boot()
+    public function boot(): void
     {
-        //
-
         parent::boot();
     }
 
@@ -41,7 +39,7 @@ class RouteServiceProvider extends ServiceProvider
      *
      * @return void
      */
-    public function map()
+    public function map(): void
     {
         $this->mapApiRoutes();
         $this->mapWebRoutes();
@@ -56,7 +54,7 @@ class RouteServiceProvider extends ServiceProvider
      *
      * @return void
      */
-    protected function mapWebRoutes()
+    protected function mapWebRoutes(): void
     {
         Route::middleware('web')
              ->namespace($this->namespace)
@@ -70,7 +68,7 @@ class RouteServiceProvider extends ServiceProvider
      *
      * @return void
      */
-    protected function mapApiRoutes()
+    protected function mapApiRoutes(): void
     {
         Route::prefix('api')
              ->middleware('api')
@@ -78,14 +76,14 @@ class RouteServiceProvider extends ServiceProvider
              ->group(base_path('routes/api.php'));
     }
 
-    protected function mapAdminRoutes()
+    protected function mapAdminRoutes(): void
     {
         Route::namespace($this->namespace)
             ->middleware(['web','check_admin_login'])
              ->group(base_path('routes/route_admin.php'));
     }
 
-    protected function mapUserRoutes()
+    protected function mapUserRoutes(): void
     {
         Route::namespace($this->namespace)
             ->middleware(['web','check_user_login'])
