@@ -1,6 +1,8 @@
 <?php
 
-    Route::group(['prefix' => 'laravel-filemanager','middleware' => 'check_admin_login'], function () {
+use Illuminate\Support\Facades\Route;
+
+Route::group(['prefix' => 'laravel-filemanager','middleware' => 'check_admin_login'], function () {
         \UniSharp\LaravelFilemanager\Lfm::routes();
     });
 
