@@ -1,6 +1,8 @@
 <?php
 
-    Route::group(['prefix' => 'account','namespace' => 'User','middleware' => 'check_user_login'], function() {
+use Illuminate\Support\Facades\Route;
+
+Route::group(['prefix' => 'account','namespace' => 'User','middleware' => 'check_user_login'], function() {
         Route::get('','UserDashboardController@dashboard')->name('get.user.dashboard');
 
         Route::get('update-info','UserInfoController@updateInfo')->name('get.user.update_info');
